refactor: rename and move appLogic script
